# backend/docs/swagger.yaml
definitions:
...
  hosts.CreateHostRequest:
    properties:
      hostname:
        type: string
      iv:
        type: string
      keyId:
        type: integer
      label:
        type: string
      password:
        type: string
      port:
        type: integer
      username:
        type: string
    required:
    - hostname
    - iv
    - label
    - username
    type: object
  hosts.HostResponse:
    properties:
      createdAt:
        type: string
      hostname:
        type: string
      id:
        type: integer
      iv:
        type: string
      key:
        $ref: '#/definitions/keys.KeyResponse'
      keyId:
        type: integer
      label:
        type: string
      password:
        type: string
      port:
        type: integer
      updatedAt:
        type: string
      username:
        type: string
    type: object
  hosts.UpdateHostRequest:
    properties:
      hostname:
        type: string
      iv:
        type: string
      keyId:
        type: integer
      label:
        type: string
      password:
        type: string
      port:
        type: integer
      username:
        type: string
    type: object
...
  keys.KeyResponse:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      iv:
        type: string
      label:
        type: string
      passphrase:
        type: string
      updatedAt:
        type: string
      value:
        type: string
    type: object
...
  models.Error:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
      statusCode:
        type: integer
    type: object
...
info:
  contact: {}
  description: API server for FlySsh
  title: FlySsh API
  version: "1.0"
paths:
...
  /hosts:
    get:
      description: Return all SSH hosts belonging to the authenticated user, with
        keys preloaded.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/hosts.HostResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
      summary: List all hosts
      tags:
      - hosts
    post:
      consumes:
      - application/json
      description: Store an SSH host configuration. Password is a client-encrypted
        blob. A key can optionally be attached.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Host details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hosts.CreateHostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/hosts.HostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Create a new host
      tags:
      - hosts
  /hosts/{id}:
    delete:
      description: Delete an SSH host by ID. Must belong to the authenticated user.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Host ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Delete a host
      tags:
      - hosts
    get:
      description: Return an SSH host by ID with its key preloaded. Must belong to
        the authenticated user.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Host ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hosts.HostResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get a single host
      tags:
      - hosts
    put:
      consumes:
      - application/json
      description: Update fields of an existing SSH host. All fields are optional.
        If keyId is provided, the key must belong to the same user.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Host ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hosts.UpdateHostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hosts.HostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Update a host
      tags:
      - hosts
...
swagger: "2.0"